single_vs_multi_agent/requirements.txt
langsmith[pytest]>=0.3.45
python-dotenv==1.1.1
requests
rich
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
vcrpy>=7.0.0